+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - fix-bug: modify connectToDatabase
 - fix-bug: fix get posts data error
 - fix-bug: fix get posts data error 1
+- fix-bug: fix get posts data error 2